- Bit stuck on verify proof of work (see first clause of ar_http_iface_server:verify_one_thing/2).
The task is to validate the proof of work of an incoming block.  The function to validate POW is
ar_mine:validate(DataSegment, Nonce, Difficulty).  Nonce and Difficulty are available in the
incoming block shadow.  DataSegment does not seem to be available in the block shadow, or the block
generated from it.  Is that right (I'm just looking at ar_serialize).

There is a function ar_block:generate_block_data_segment/6, but it's not clear to me whether this
extracts a data segment from a block or generates a new data segment.