- Refactored ar_node:validate/8 (experimental).
I've put this in a separate branch in case people think the code is too fancy.  The changes are all
to ar_node:validate/8.

One thing I don't like about ar_node:validate/8 (because it tripped me up) is that each thing that is checked is touched four times, on four separate lines (lines nos here from before changes):

- 1. when the check is made and the variable introduced, e.g.:

    L 1625    Wallet = validate_wallet_list(WalletList),

- 2. when the ar:report list is made up

    L 1646    {block_wallet_validate, Wallet},

- 3. in the ar:d block

    L 1676    case Wallet of false -> ar:d(invalid_wallet_list); _ -> ok  end,

- 4. in the final andalso block

    L 1691    andalso Wallet

I have refactored the function so each check is mentioned only once.  The Checks list at line 1625
has for each check:

    {Label, ErrorMessage, Verification},

This list is run through a foldl at line 1672.

All but one of the checks returns a true/false, one (ar_mine:validate/3) returns false/<integer>, so the fold catches anything that is false (i.e., not false means ok).

All but one of the checks is side-effect free.  The remaining check, for the recall block, triggers
extra actions (see lines 1692f).

Is the new version better or worse in terms of clarity?

The bad news is that this refactor causes a huge amount of damage to the make tnt tests:

    Before:
    =======================================================
      Failed: 3.  Skipped: 0.  Passed: 80.
    =======================================================

    After:
    =======================================================
      Failed: 27.  Skipped: 0.  Passed: 56.
    =======================================================

If people think these changes are sensible, I'll track down why those breakages are happening.
Otherwise I'll drop it.