- post_block_to_unjoined_node_test passing *in shell only*.
This test passes in the shell because the node at localhost is not joined
to a network.  The test fails in `make tnt`, because the localhost node is
joined to a network.

The json sent in the test is not a valid block.  I have added a check for this,
and the server returns 400 with body "Invalid block."

n.b.: the handle POST /block uses a verifier function rather than nested cases.