Fix /unsigned_tx requiring id, owner, signature
When posting an unsigned TX, the id, owner and signature should be
generated on the node and not be posted by the client. This fixes a
bug where all three are required to be set by the client (which later
will be overwritten when signing the transaction). This commit fixes
this bug and also the bad test.