Refactor ar_meta_db as a gen_server
I had some problems getting all tests to pass.
This is because some module tests are dependent on each other.

I tried many different strategies but the least invasive was to
move the module test for ar_meta_db. Now it is the last in the list.
Otherwise I had problems with ar_storage and ar_http_iface_tests.