Respect Wildfire score when fetching recall blocks
A peer used to be picked randomly. As a result, a node was likely to
pick a slow peer or even an unavailable one (e.g. a node in the
polling mode), what lead to bad fork recovery times.

This also changes the behaviour for fetching blocks when joining and
in fill_to_capacity (ar_join).