Use the correct timestamp for calculating the diff
The diff for the next block should be calculated with the same
timestamp used for the next block. Currently, the next_diff/1 function
is using os:system_time/1 which in theory could be different from the
block timestamp. In reality, the risk of returning wrong difficulty is
super slim. This improvment will make more sense later when we update
the block timestamp while mining.