Create a distributed setup & test storage firewall
The distributed setup is motivated by the need to test the ability of a node to process blocks with the transactions that do not pass its firewall. We launch two nodes with isolated storage, the firewall is set up for the first node so that it verifiably does not store data. The second node is used to fetch the data for validation.