RandomX NIF: Use Dirty schedulers instead of threads
Make each NIF call Dirty (see "Dirty NIF" in
http://erlang.org/doc/man/erl_nif.html) instead of doing the heavy
work in a thread. This simplifies the code while mostly keeping the
same behavior (and performance).

There is only one dirty scheduler per regular scheduler. If all the
dirty schedulers are busy, then the next Dirty NIF call will be
blocked. This will cause a delay of hashing if running on a
single-core machine and the only dirty scheduler is busy generating a
dataset. This could delay verifying new blocks received but should be
fine.

Co-authored-by: Martin Torhage <martin@arweave.org>