Implement TX replay protection via block anchors
The rationale is to support multiple transactions per wallet per block,
avoid replay attacks while not requiring nodes to store the full weave,
and support out of order transaction propagation. The solution also
attempts to pick a reasonable tradeoff between memory pool abuse
and the probability of network consensus on which transactions to mine.

Every valid transaction now has either a "wallet list anchor" or a "block anchor."

A wallet list anchor is what every valid transaction before the fork 1.8 has -
a reference to the last transaction of the wallet. It is still possible to create
such transactions after fork 1.8, but only one such transaction per wallet per block is allowed.

A block anchor can be attached to transactions after fork 1.8.
For backward-compatibility reasons, the same "last_tx" field is used to attach the anchor.
A valid block anchor is an independent hash of one of the last 50 blocks of the weave.
It is wise to use an anchor about 25 blocks old. If the block is too fresh,
it is more likely to be orphaned. If it is too old, it might not have enough time
to be included in the weave due to the network propagation delay.

After the fork, nodes do not accept transactions extending the 250 MB memory pool limit.

When nodes pick which transactions to include into a block, they choose
the biggest by size transactions first. This approach makes memory pool abuse more expensive.
If two transactions are of the same size, nodes choose the one with an older block anchor
since it has less time left until it becomes invalid.

It is still not possible to spend tokens sent by transactions from the mempool
to ensure smooth out of order propagation across the network. For the same reason,
if there are multiple transactions sending tokens to a new wallet, every transaction
must include the wallet generation fee.

Memory pool can potentially contain transactions which conflict by sharing the wallet
list anchor or by spending funds not sufficient for both transactions.
Allowing for conflicting transactions in the mempool is essential for the network consensus.
Miners do not know which transactions other miners have. They need to
keep potentially valid transactions to be able to verify new blocks from other miners.

The commit also switches `GET /wallet/[Addr]/txs(/[EarliestTXID])` to use ArQL instead
of the wallet list since the wallet list does not include block anchored transactions any longer.