Reduce truncated files on write
Since we don't do atomic file writes in ar_storage, if the filesystem
is full we might write empty JSON files. This will show up as errors
when trying to JSON parse them later.

This adds a bit of a margin to the disk_space metric for filesystem
overhead, so that we hopefully stop writing just before the filesystem
really is full.