Fix Mnesia startup after bad interrupt and hostname change
If a hostname changes and a transaction gets interrupted by a node shutdown, Mnesia refuses to pick up the state assuming a distributed mode with another hostname responsivle for resolving the current state. Since we are only using one Mnesia node all the time, we can force it to load the table.