Improve efficiency of ArQL index update
The speedup is achieved by storing block height and block hash in a separate type=set table. This frees us from the need to remove old entries from the arql_tag table on every write (because the same transaction can be included in a different block after switching to a new fork). The solution is fully backwards-compatible since both tables are checked on every read (an overhead added to the previous reads is little).