Pass block shadows to new_block ADT callbacks
The new_block event occurs when a new block arrives and passes
edge verification (the timestamp and PoW checks). The payload
of the event is the block shadow.

The commit removes the code that attempts to search for the
transactions of the block on disk. This is problematic for
two reasons:

1) There are very likely no transactions on disk at this point.
   Even if they were on disk, they could be missing due to
   the propagation issues or for being rejected by the firewall.
2) Due to the potentially big transaction size, including
   transaction payloads in the events is problematic.
   Specifically, webhooks cannot expect the receiving side to
   accept an arbitrary big payload - some popular services have
   a 10 MB HTTP body limit. Therefore, we expect users to fetch
   transactions by ID on demand.