FIXUP make fork recovery work together with 2.0 transition
The commit includes the following changes:

1. Block struct's block_index field was renamed back to hash_list.

Blocks do not need to carry the whole block index with them. When a block
shadow is distributed, the field only needs to contain the last 50 hashes
for the receiver to be able to reconcile it with the state. The Merkle root
of a hash list is also built off of the hashes only.

2. A "legacy hash list" is memorized by the node when it transitions to 2.0.

This hash list contains the hashes of all the blocks up to the 2.0 block (exclusive).
It is used for identifying the diverged hashes if a fork occurs before 2.0 and grows
past 2.0. Every time the node transitions to 2.0 the new legacy hash list is stored
and the checkoint is updated.

3. The fork recovery code is simplified.

The reconstruction of the diverged hashes is improved, previously it was spread
around multiple places, some logic was redundant.

4. The node now downloads 100 blocks on join.

This is made for the sake of registering the recent 100 block anchors in the state,
what allows to further simplify and speed up the fork recovery process, which
previously included downloading the missing block anchors from the trusted peers.

5. Nodes do not reject height + 1 blocks with illicit transactions any longer.

The local firewall still protects the storage though. The removed logic does not
make sense for the new protocol, which makes it unrealistic to download and scan
data "on the fly".

6. The GET /block/hash/[hash]/wallet_list endpoint was simplified to not depend on the block index format.

It is required so that nodes can fetch wallet lists for blocks by their v1 hashes.

Some notes about tests:

In ar_multiple_txs_per_wallet_tests:joins_network_successfully_test_, a node
transitions to 2.0, another node joins it, the original node forks off before 2.0,
the second node recovers via the second 2.0 transition.

In ar_multiple_txs_per_wallet_tests:recovers_from_forks, a node before 2.0 fork recovers to
a node after 2.0 and experiences its first 2.0 transition in the middle of the fork recovery process.