Upgrade RandomX to v1.1.7
v1.1.7 is the latest version of RandomX, which contains major performance improvements (up to 20%).

The commit includes a test for the RandomX compatibility, which can be used in the future
compatible upgrades.

bulk_hash_fast_nif was switched to the new RandomX API, functions calculate_hash_first,
calculate_hash_next, and calculate_hash_last, which provide the main perfomance increase.