FIXUP include PoA inside PoW
Proof of access needs to be inside proof of work so that it can't
be "front runned".

In addition to that, the commit contains a few major improvements:

1. After 2.0 the block data segment construction is part of the independent hash construction.

This reduces the computation burden and makes it much easier to maintain the block hashing mechanism.

2. The new block data segment construction algorithm allows the miner to refresh data segment during mining quicker.

The miner's reward wallet is hashed in a separate step instead of rehashing the whole wallet list every so often.

3. The hash list merkle root is turned into a block index merkle root, as it has to contain weave sizes after 2.0.

4. The miner does not validate its own blocks and does not do other redundant work after mining a block.

Blocks are taken as they are returned from ar_mine instead of being constructed again in ar_node_worker (ar_weave:add is deprecated and
only used in tests now).