FIXUP double-hash to prevent length extension atttacks
The attack can be used to make someone believe a set of invalid
transaction identifiers belongs to a block. An attacker can come up
with some garbage, which hashes to the same tree node when combined
with the transaction size. Hashing the leaf node before combining it with
the size addresses the issue. In addition to that, the commit introduces
an extra hashing iteration at every level of the tree construction
to stay on the safe side.