Remove redundant wallet list hash reconstruction
The redundancy is removed via storing the hash in #block.wallet_list_hash upon
initial reconstruction.

write_wallet_list/1 is not exported anymore as it is only used internally.

write_block/2 was added which can be used to store a block header without storing
the corresponding wallet list. The motivation is to speed up the 2.0 transition.