Do not attach hash lists to blocks
The only two cases where we need block's previous hashes is when we gossip
a new block to peers and when we receive a block from peers. Even then,
we only need the last ?STORE_BLOCKS_BEHIND_CURRENT hashes.

In 2.0 we do not need the complete hash list for generating a data segment,
we use its merkle root instead.

To determine the current recall byte, we use the block index, which is passed
separately.

Therefore, no reason is left for blocks to carry huge hash lists around.