Add support for syncing large data
== Motivation ==

The motivation is to sync arbitrarily large portions of the weave efficiently and in a way that keeps the data distribution across the network uniform, while making the storage and processing overhead grow only with the size of the data stored, not with the size of the whole weave.

The syncing needs to be efficient in the presence of chain reorganizations taken a significant amount of data may become orphaned. Also, we do not want to sync the data twice if it gets orphaned and then reincluded into the weave.

Finally, we want to provide a good user experience. Users need to be able to upload some data before the corresponding transaction is confirmed. Also users need to seed data to nodes without knowing the block it is included in.

== Design Decisions ==

- The chunks are indexed by their global end offset on the (0, weave size] half-interval. It allows nodes to keep a compact representation of what is synced - a number of non-overlapping intervals, see the ar_intervals module. The number of chunks in the weave can grow arbitrarily big thefore an alternative of indexing chunks by their ID would not allow for efficient data discovery.
- Nodes share the information about which portions of the weave they have. A node only attempts to fetch a chunk from a peer if
the chunks's offset is covered by the peer's "sync record". Nodes pick chunks randomly so that the data distribution is uniform. Yet, storing too many non-overlapping
intervals results in a big sync record which is expensive to transfer, therefore nodes aim for syncing continuous patches of the weave so that the total number of
intervals they have stays under 10000 (~800 KiB in Erlang Term Format). If they happen to have more intervals, a random subset of intervals is shared with peers.
- The information about the recent chunks is stored in a separate index, a "disk pool". When a reorg occurs, a lot of data is potentially orphaned - it would be too expensive to move it around. Therefore, the only thing we do on reorg is a range delete of the chunks from the offset index. A background process which traverses the disk pool in the background picks up the orphaned chunks and returns them back into the offset index. The files with the chunks themselves are only written once when the chunk is seen the first time. The same process also removes the chunks from the disk pool and from disk if they expire. The disk pool also stores pending chunks and treats them much like the orphaned chunks - picking them up if they get confirmed or removing them by expiration.

== API ==

The existing mechanism of syncing v2 data is kept intact, to ensure a smooth transition.

GET /tx/<id>/data continues to serve data smaller than 12 MiB. For bigger transactions, a new GET /tx/<id>/offset endpoint is introduced, which tells the transaction dimensions - the client can use it to learn chunk offsets and fetch the transaction chunks via GET /chunk/<offset>.

POST /tx still accepts data, so it is compatible with existing clients. To submit a larger transaction, the new client needs to seed it in chunks using a new POST /chunk endpoint.

== Configuration ==

- disk_pool_data_root_expiration_time, the time in seconds of how long a pending or orphaned data root is kept in disk pool, 2 hours by default;
- max_disk_pool_buffer_mb, the maximal total size in mebibytes of the pending chunks in the disk pool, 2 GiB by default;
- max_disk_pool_data_root_buffer_mb, the maximal size in mebibytes of the pending chunks in the disk pool, per data root, default is 50 Mib.

== Metrics ==

- v2_index_data_size, gauge, the total size of the chunks stored, in combination with the weave_size metric can be used for monitoring a percentage of the weave synced;
- disk_pool_processed_chunks, counter, the running counter of processed disk pool items, its change rate is indicative of how much time is spent processing the disk pool (scanning for expiration, picking up pending and orphaned chunks, removing orphaned chunks) and how fast it is processed;
- disk_pool_chunks_count, gauge, the number of items in the disk pool, together with disk_pool_processed_chunks helps to asses how clogged the disk pool is;
- pending_chunks_size, gauge, the total size of the pending chunks in the disk pool.