Push attached v2 data to trusted peers
Sending data along with the transaction header is a more convenient but less reliable
way to seed data to the network, compared to submitting data with proofs chunk by chunk
to several nodes until a desired level of initial redundancy is achieved.

The commit strives to strike a balance between overloading the network with too much
gossip (of attached v2 data) and the risk of failing to seed the data at all.