Do not prioritize new txs over those moved to pool
The motivation for such prioritization was to reduce the amount of transactions
the network has to come to consensus with during high load. However, it comes at the cost
of reducing the transaction write rate during these times. The recent improvements in
transaction processing and handling missing block transactions motivate us to roll it back
and allow the nodes to reach their full capacity.