Turn off HTML browsing by default
In many countries and regions, content censorship exists. Some illegal content may cause the node to be shut down and require the owner to take legal responsibility. Therefore, it is recommended to turn off HTML browsing by default.
Although this PR cannot solve the problem from the root cause, it is currently the easiest way to deal with it.
The transmission of AR content is direct and plain text, and it may also be risky if the review agency wants to do it. It just looks fine for now. May need to do content encryption or use secure communication like HTTPS.