Drop the automatic_rocksdb_repair flag
The unflushed column families are lost during the repairing process (https://github.com/facebook/rocksdb/wiki/RocksDB-Repairer). The workaround we used to have does not always work - an attempt to flush a column family when starting after an emergency shutdown may hang.

In the unlikely event of a disk corruption, the repairing may be attempted via the RocksDB CLI tool.