Do not consider single small chunks unattractive
Chunks which are the only chunks in their merkle trees can be produced at a rate of at most 1k/block.
Moreover, the corresponding transactions are less likely to pass the distribution priority queue
because they are small and cheap. Therefore, they are not as troublesome to process as are small
chunks of bigger txs.