Fix public bit size in RSA PSS signature check
From @onesk:

Adjusts the PSS padding to correct length in case modulus bit length is less than 4096 - the byte size should be rounded up, not down.
A test case for this is evidenced in #283. Arweave-js would succesfully verify said signature via crypto.subtle, but Erlang implementation fails to do so. Not only it rejects the perfectly valid signature, it also crashes the verification process.