Reimplement the first test using Common Test
Common Test is a builtin framework well suitable for writing black box tests that are currently
written with the help of home-written helpers and eunit.

The framework takes care of setup and teardown what simplifies adding new test scenarios.

The test/ar_test_lib.erl module is introduced as a replacement for test/ar_test_node.erl.

The planned next steps are:

- add scripts and helpers to make it easy to run and debug individual tests, such as
  a script for tailing nodes' logs, an Erlang shell function for running individual tests,
  a mechanism for creating CT breakpoints to investigate nodes' states after a test error;
- migrate all the helpers to test/ar_test_lib.erl;
- re-implement all black box tests using Common Test;
- include new tests in the CI pipeline.

Co-authored-by: Lev Berman <ldmberman@protonmail.com>