mention disk_space param in "out of space" message
The current message is confusing when there is enough disk space in general, but the used space is below `disk_space` limit CLI parameter. 

This PR adds an explicit mention of `disk_space` param to prevent confusion for new users e.g. if that param is incorrectly set or if the parameter limit is the issue instead of overall available disk space.