Add support for syncing chain headers in parallel
The pace is controlled by the header_sync_jobs command line and config file parameter.

Syncing historical headers is a slow process mostly because there are simply too many historical
blocks and transactions to download, the serialization format is not very compact, and API does not
batch them. The commit provides a very straightforward and simple way to speed it up.