Add optional support for address checksums
Make the node recognize a new optional wallet address encoding:

[Base64Url encoded address]:[Base64Url encoded CRC-32 checksum]

The motivation is to avoid incidentally sending tokens to the wrong address. The upgraded nodes
will recognize the new format and validate the checksum. For backwards-compatibility, the upgraded
nodes will keep serving and gossiping the old format. Once the ecosystem adopts the new format,
we can ship an upgrade where the nodes also serve and gossip the new format. Once the network
adopts this, we can drop support for the old format.