Better determine when to pause disk pool scanning
It is important to process the disk pool as fast as possible and at the
same time avoid wasting disk IO on yet pending or already indexed and
waiting for more confirmations chunks. The commit measures the time of
the last full scan - starts a new scan if it took longer and puts
scanning on hold for some time otherwise.