Add basic support for blacklisting byte ranges
The byte ranges can be specified as "[start],[end]" lines in the transaction
blacklist files and transaction blacklist endpoint responses.

Whitelisting a TXID does not save its byte range from blacklisting if a
separate range intersecting the transaction's range is specified.

The ranges specified in the whitelist files/responses are ignored.

The specified end offset should match the end offset of a chunk.