Cross-sync/repack data across storage modules
Only attempt to sync the data from the network after copying/repacking
the data across the intersecting storage modules.

The data is also fetched from the default (2.5) storage. The 2.5 -> 2.6
repacking is performed via the new mechanism too, what should speed
it up.

Make the data chunk cache and packing cache configurable, set the
defaults based on the estimated available for allocation amount of free
memory.

Repair the broken records in the process.