Cache the longest chain (#220)
* First pass at a cache

* fix up some ets mistakes

* Better, atomic cache updates

* Complete testing of longest_chain cache

* add some more output to make sure we're rebuilding the cache at the right times

* Remove the race condition test. It wasn't testing the correct thing.
Since we don't read from the cache in order to update it there isn't
really a cache-specific race condition to test. What we can observe, though,
is if multiple nodes come in at the same time - this can cause the
get_longest_chain_block_txs_pairs to give different results based
on exactly when each node is processed and exactly when the function
call starts and ends. However this lack of determinism isn't specific
to the cache and, arguably, isn't a bug to fix as there is some
legitimate ambiguity about what the longest chain is at the moment
that it is being concurrently updated. That ambiguity is quickly resolved
as the concurrent updates complete.

Long story short: I'm removing the test as it wasn't testing the
the new cache behavior, and may not have been asserting the right
condition anyways.

* remove debugging output