N.2.6.8.old (#424)
Release 2.6.8
Release Number 63

This release introduces a patch that adds to the transition period before the
activation of Arweave 2.6’s trustless price oracle, in order to give miners
additional time to on-board packed data to the network. The release delays
the onset of the transition window by 4 months, and extends the interpolation
between of old and new pricing systems to 18 months, from 12.

This release introduces a hard fork that activates at height 1,189,560,
approximately 2023-05-30 16:00 UTC. You will need to make sure you have
upgraded your miner before this time to connect to the network.