Ensure that the VDF client only processes each VDF step once, even (#459)
* Ensure that the VDF client only processes each VDF step once, even
if the server sends the same step multiple times.

Specifically: handle the case where the Server has been pushing steps
for a "stale" SessionKey (i.e. the entropy reset line has passed, but
the old SessionKey continues to be used until the VDF server processes
a new block).

This overlap in session caching is intentional.  Before a new block is
found steps keep being written to the previous session (identified by
{NextSeed, IntervalNumber}) and once a block is found, the steps above
the reset line, if any, are added to the new session (with the new NextSeed).

The intention of the overlap is to quickly access the steps when validating
B1 -> reset line -> B2 given the current fork of B1 -> B2' -> reset line ->
B3 i.e. we can query all steps by B1.next_seed even though on our fork the
reset line determined a different next_seed for the latest session.