Initialize ar_poller.in_sync_trusted_peers with the set of trusted peers (#464)
* Initialize ar_poller.in_sync_trusted_peers with the set of
trusted peers so that we only remove a peer when it is truly out
of sync. Previously the set was initialized empty which meant
a peer was only added if it went out of sync (due to the 5 minute
timeout). This created a situation where a singl out of sync peer
could cause the "Out of sync" console message to be printed.

By initializing the set with all trusted peers, the message is
only printed when all peers go out of sync

* Change the testnet network to arweave.2.7.testnet so that
old nodes from the 2.6 testnet days don't accidentally interfere

* Allow testnet nodes to cleanup up orphaned data