Have ar_http_iface_client::get_info() use a map
Previously it manually built a JSONStruct/PropList and used that
intermediate format to store/access individual info items. However
we'll soon be adding recent and forks keys which will have more
complex structures. Manually building and searching a proplist will
get hairy. This commit moves to using an erlang map as the
intermediate format and relying on the jiffy encode/decode funcitons
for converting to/from json (rather than doing it by hand).