Do not adjust the hash rate in the reward history
Adjusting the hash rate by (TwoChunkCount + OneChunkCount) / TwoChunkCount counts
the number of all the hashing attempts. In other words, the adjusted value is useful
when we want to see the total amount of CPU work put into mining a block.
This is not what we use it for. We use it as a denominator when computing
a share contributed by a single partition - see ar_pricing:get_v2_price_per_gib_minute.
Therefore, the hash rate computed here needs to have the same "units" as
the hash rate we estimate for the partition - the "normalized" hash rate
where a recall range only produces 4 nonces from one recall range (chunk-1)
plus up to 400 nonces (chunk-2).

Note that we did not even adjust it by (TwoChunkCount + OneChunkCount) / TwoChunkCount but
by 100 div (100 + 1), what is wrong.