Arweave Should Not Build Before Being Stopped
This commit fixes an issue related to `./bin/stop` script and
`./bin/arweave.env`. When stopping Arweave application in
dev environment, a full release is always created. It is not
clean and could probably leads to errors (even if was not the
case previously).

The compilation is now made only when SCRIPT_ACTION variable
is not set to "stop" or if `./bin/arweave-dev` does not exist.