fix: get ar_tx_blacklist_tests to pass
Prior to this change ar_tx_blacklist_tests relied on every test node
having pack_served_chunks set and being able to repack chunks on
demand. To preserve that behavior we have to add each node to
the local_peers config for the test nodes.

Note: I haven't run this change on other tests yet, it's possible
it causes something else to fail (in which case we may want to
clear local_peers in that failing test)