Change Logging Feature Mode
This commit creates a new module called `ar_logger` where all
functions related to logging are present. The goal here is
also to get rid of the `disk_log` logging type causing sync
issues on rsyslog and using a more conventional "Unix way"
to store logs.

The previous method using `disk_log` was using a wrap method,
creating also two files ending with .siz and .idx to follow
the logs. One of the important feature of `disk_log` is to
repair logs, but it seems this feature is not used. Instead,
using a simple append method with `logger_std_h` mode will
make things easier when collecting logs.