Fix High Load Average on Transaction Interface
When a server has the full transaction locally, and uses
spinning disk, the amount of requests from other nodes is
too high, and leads to a bottleneck on rocksdb database.

This commit adds a semaphore to control the amount of
requests on `ar_http_iface_middleware:handle_get_tx/3` function,
to avoid a node to be overflooded by requests.