Optimize and Improve Github CI/CD
Github Workflow has been improved and optimized to run in a parallel
environment using dynamic runner spawned on demand. Artifacts are now
uploaded and copied across each builds. A build will always start from
a fresh installation (meaning the compilation time will take longer).

This commit fixes also few bugs in the test suite, causing random
crash during execution. They are mainly related to the load of the
system where the tests are running.

 - A race condition was found when starting test on test
   peer. Sometimes, peers are not ready to start the tests and crash.

 - Many race conditions because of timeouts, mainly in
   `ar_vdf_server_tests`, `ar_http_iface_tests`, `ar_poa_tests`
   and `ar_tx_blacklist_tests`

The number of workers has been set to `8` instead of `12`. The
current server is having trouble dealing with more than `12`
workers in parallel.

A cache has been added for dependencies, this is inefficient to
fetch dependencies for every build. To avoid that, deps are
updated when rebar.lock checksum is different from the previous
build. If deps cache is not found on the cache store, it is
created and uploaded.

see: https://github.com/ArweaveTeam/infra/issues/114