fix: reduce impact of entropy overhang + fix some bugs
We always have to generate more entropy than there is data in a
partition. Partly because a partition at 3.6TB is not evenly
divided into chunks (at 256 KiB each), and also partly because of
how entropy is sliced up and allocated to chunks.

This extra entropy is called the "overhang". Previously the overhang
would stack - i.e. the overhang on partition 0 would actually be
applied to the first bit of partition 1, and so on. By the time
we got to a hgher partition, a significant fraction of the chunks
in the partition would rely on entropy that belons to the previous
partition.

This made accounting for entropy partitions vs. recall partitions
a little confusing, and it also had a negative impact on performance
when packing non-consecutive partition.

This change localizes all entropy to a single partition - basically
discarding the overhang rather than applying it to the next partition.

at 256 MiB the amount of entropy discarded per partition is relatively
small. The benefit of this change is that there should no longer be a
negative performance impact to packing non-consecutive partition. This
impact was particularly noticeable when packing non-consecutive higher
partitions (e.g. packing just partition 30 would suffer from
a multi-hour period during which entropy generation was very slow due
to the early chunks in the partition relying on entropy that was
predominantly used to pack chunks for partition 29).

In addition to the overhang change, this commit addresses a couple
of bugs in the prepartion and enciphering phases