Improve Domain Name Resolution
Current arweave behavior is to select one IP address when
a record contains more than one value (round robin). The
new behavior is to retrieve the complete list and then,
give the possibility to create DNS pool. Each IP address
is simply added in the peer list (with the default port
if not specified).