Improve Domain Name Resolution
Current arweave behavior is to select one IP address when
a record contains more than one value (round robin). The
new behavior is to retrieve the complete list and then,
give the possibility to create DNS pool. Each IP address
is simply added in the peer list (with the default port
if not specified).

Few DNS pools have been created to help users and miners to select the closest trusted peers for their region. It has been inspired by the [`pool.ntp.org`](https://www.ntppool.org/zone) convention. Below the list of the current domain name pool available:

 - `pool.arweave.xyz`: contain all available trusted peers
   - `asia.pool.arweave.xyz`:
   - `europe.pool.arweave.xyz`
   - `north-america.pool.arweave.xyz`
   - `india.pool.arweave.xyz`
   - `oceania.pool.arweave.xyz`

The idea is to offer a simple naming convention hiding the name of arweave hosts and their IPs. In case of change, users will always have more than 1 node available from the pool. It will help system administrators to deal with the outages.

Furthermore, it will also be easier to monitor and offers a better view for everybody. Most of our trusted peers are still publicly available and a [status page](https://status.arweave.xyz/) for each of them can be seen. This is a quick view to check if the peer is correctly to HTTP requests.

This new implementation is quite easy at first glance, it only replace `inet:getaddr/2` by `inet:getaddrs/2`. This simple change allows to get the full list of IP addresses set on a domain name instead of one randomly. This function is called when a peer is configured via `ar` module (from command line or configuration file).

These peers (from the DNS pool) are then available in the trusted peers list through `ar_peers:get_trusted_peers/0`. When setting `peer europe.pool.arweave.xyz`, the function returns 6 peers available:

    ar_peers:get_trusted_peers().
    % returns:
    % [{91,239,53,95,1984},
    %  {57,128,189,200,1984},
    %  {67,220,66,3,1984},
    %  {57,129,64,100,1984},
    %  {23,88,18,239,1984},
    %  {23,166,88,225,1984}]

To be sure, `dig` command can be used to list the whole list IP addresses available from this record. In this case, dig also returns 6 peers, with the same IP addresses.

    $ dig +short europe.pool.arweave.xyz
    23.88.18.239
    57.129.64.100
    23.166.88.225
    91.239.53.95
    67.220.66.3
    57.128.189.200

Let check now what happens when a node goes down. Instead of shutdown the node directly, a new iptables rule will be added:

    # drop all packets coming from 23.88.18.239
    sudo iptables -I INPUT -i ${interface} -s 23.88.18.239 -j DROP

    # drop all packets going to 23.88.18.239
    sudo iptables -I INPUT -i ${interface} -d 23.88.18.239 -j DROP

    # check if the peer is still reachable
    nc -zvw3  23.88.18.239 1984
    # it should return:
    # nc: connect to 23.88.18.239 port 1984 (tcp) timed out: Operation now in progress

    # kill the remaining connections if present
    sudo ss -K dst 23.88.18.239
    sudo ss -K src 23.88.18.239

When starting the server:

    Setting the data chunk cache size limit to 1100 chunks.

        Peer 23.88.18.239:1984 is not available.

    Joining the Arweave network...
    Downloaded the block index successfully.

All other nodes are still reachable, only the blocked one is not available.