Shutdown Procedure for remaining sockets
When stopping a node, cowboy is waiting for all connections to
be correctly closed. Usually, it waits for the buffer to be
empty before closing the socket. In most of the cases, this
behavior is acceptable, but in some cases, in particular when a
connection is very slow, this can have a direct impact on the
delay to shutdown a node.

This commit adds a shutdown procedure when stopping the client
connections. It has 3 stages: (1) when a node is stopped, all
sockets are set in read-only mode and all clients are noticed of
this change, and they will try to fetch the data before closing
the connection (2) if some nodes are still present, the procedure
sends another closing requests and will wait for a specific delay
(3) if the connection is still up, the connection is simply killed,
linger option is set to {true, 0} and the socket is definitively
closed.

This draining procedure was inspired by ranch documentation example,
that can be found at: https://ninenines.eu/docs/en/ranch/2.2/guide/connection_draining/

The timeout can be set with shutdown_tcp_connection_timeout
parameter. Its default value is 60s.

see: https://github.com/ArweaveTeam/arweave-dev/issues/817