Shutdown Procedure for remaining sockets (#745)
When stopping a node, cowboy is waiting for all connections to be
correctly closed. Usually, it waits for the buffer to be empty before
closing the socket. In most of the cases, this behavior is acceptable,
but in some cases, in particular when a connection is very slow, this
can have a direct impact on the delay to shutdown a node.

This commit adds a shutdown procedure when stopping the client
connections. It has 3 stages: (1) when a node is stopped, all sockets
are set in read-only mode and all clients are noticed of this change,
and they will try to fetch the data before closing the connection (2)
if the connection is still up, the connection is simply killed, linger
option is set to {true, 0} and the socket is definitively closed. The
last timeout explicitly kill the process and the port to free the vm.

Furthermore, ar_http_iface_server is now an interface to cowboy and
uses gen_server behavior. By default it will catch shutdown messages
and start the procedure to clean all the connections.

This draining procedure was inspired by ranch documentation example,
that can be found at:
https://ninenines.eu/docs/en/ranch/2.2/guide/connection_draining/

A quick refact of ar_node_sup and ar_node_worker modules have been
done. ar_node_worker was in charge to start/init other workers, this
was not the right place. Those workers are now managed by ar_node_sup.

The timeout can be set with shutdown_tcp_connection_timeout
parameter. Its default value is 60 seconds. The default shutdown
timeout for the supervisor has been set to 300 seconds (5 minutes).

Fix an issue with ar_semaphore

see: https://github.com/ArweaveTeam/arweave-dev/issues/817

Co-authored-by: arweave <arweave@mathieu.arweave.xyz>