feat: expose repack_cache_size_mb as a configuration option
There are 4 values which control repack-in-place performance
and resource use:

- exposed: repack_batch_size: number of consecutive chunks to read at a time
- exposed: repack_cache_size_mb: maximum size of the repack chunk cache
- internal: write_batch_size: number of chunks to write at a time.
                              hardcoded to 1024
- internal: num_entropy_offsets: how much entropy to use at once

By default when we generate entropy we generate 1024 chunks worth.
But this creates some memory pressure since if we want a repack_batch_size
of 100 (for example) that means we need to cache 100 different entropy
footprints for a total size of 102,400 chunks, or ~25GiB. To address this
we can set the num_entropy_offsets less than 1024. This means we will
generate some entropy and then throw it away. For example if
num_entropy_offsets is 128, we only use 128 chunks of entropy each time
instead of the full 1024, and so will end up calculating the same entropy
8 times. Then node will set num_entropy_offsets automatically in order
to satisfy the repack_batch_size and repack_cache_size_mb.

This can be a useful config option to balance RAM vs. Disk IO vs. CPU