Improving Shutdown TCP Connection Procedure
This commit try to find a way to close HTTP/1.1
connections with high latency peer. Indeed, HTTP
protocol is half-duplex (at least for the version
used right now) and some peers can request huge
amount of data with a poor connection.

In this situation, the server is stuck with TCP
waiting for an ACK, even when sending a FIN for
the shutdown procedure. A similar behavior seems
to be applied using close and linger. Still unclear
if it's a bug on Erlang side or some behavior
from arweave client/server configuration.

The new shutdown procedure is more brutal, in both
soft and hard method. The list of sockets are
directly listed from erlang:ports/0 functions and
filtered to extract only the sockets to close. Then,
the code loops until all sockets are closed. This
procedure is now executed as prep_stop callback
in ar application.

A new module/process called ar_shutdown_manager has
been created and added on the arweave application tree.
This process should be started first and stopped last,
it is in charge of cleaning up the remaining connections
from gun and cowboy when the application is going down.
Every shutdown procedures and related functions have been
migrated to this module.

Many socket parameters have been modified to improve
the closing time of the socket. Users can also tune
their system if needed, in particular these values on
Linux, by using sysctl:

    net.core.default_qdisc
    net.ipv4.tcp_congestion_control
    net.ipv4.tcp_dsack
    net.ipv4.tcp_fack
    net.ipv4.tcp_fin_timeout
    net.ipv4.tcp_keepalive_intvl
    net.ipv4.tcp_keepalive_probes
    net.ipv4.tcp_keepalive_time
    net.ipv4.tcp_min_rtt_wlen
    net.ipv4.tcp_orphan_retries
    net.ipv4.tcp_retries1
    net.ipv4.tcp_retries2
    net.ipv4.tcp_sack
    net.ipv4.tcp_synack_retries
    net.ipv4.tcp_syn_retries
    net.ipv4.tcp_tw_reuse

More documentation is required on this part, to
explain how to configure correctly these values
depending of user's needs.

In some specific case, the application can
still hang a bit during shutdown. The active connections
can be listed using lsof or ss commands.

    ss -p sport = 1984

Where 1984 is the port used by arweave server. This
command will list all connections with their state on
the process using them. To kill them manually - only on
Linux - the flag -K can be added.

    ss -pK sport = 1984

This is not a clean way of closing sockets, but if the
service got stuck because of a remaining connections,
it can help.

see: https://github.com/ArweaveTeam/arweave-dev/issues/859