Fix Timeout Issue
erlang:halt/0 or erlang:halt/1 should not be used,
or only if something really bad happens. usually,
when an erlang application must be stopped, even
more when it's a release, init:halt/0 or init:halt/1
should be used instead. init:halt functions will
take care to close everything before the shutdown.
Using erlang:halt functions are doing the opposite,
and generate a core dump, without cleaning anything,
it could also explain why many time the peers are in
bad state after a shutdown.

If only a few part of the trusted peers have a bad
clock, it should not be a problem. The previous
method was to stop arweave when one peers had
some trouble, if more than 10 nodes are configured
(for example peers.arweave.xyz), then it should not
be a problem.

see: https://github.com/ArweaveTeam/arweave-dev/issues/907