Improve service state
ar_shutdown_manager can now set the state of the system as running
or shutdown. When the system is in shutdown state, part of the
application should not be used, for example, no new connections
from the peer can be created, and no new timer interval can be
created.

see: https://github.com/ArweaveTeam/arweave-dev/issues/915