OpenSSL Lite Integration
OpenSSL 1.1.1 was deprecated months ago, but
Arweave still depend on it for SHA256 functions.
This commit integrates a stripped version of
OpenSSL 1.1.1 with only SHA256 functions and few
other mandatory requirements. The code is compiled
and then statically linked into the NIF library.

Compiled and tested on Ubuntu 22.04 with Erlang R24.
Compiled and tested on MacOS 15 with Erlang R24.

see: https://github.com/ArweaveTeam/arweave-dev/issues/916