Fix absolute symlink issues on standalone runners
When building arweave with rebar3, rebar3 generates
many absolute symlink. If the build has been done
on user-A, the link will point to user-A working
directory. Now, if a new job is started with the
same artifact, but on user-B, then, the symlinks
will still point to user-A working directory. In
this situation, user-B does not have access to the
content of the directory, and then, it fails.

To avoid this issue, we can create a shared
directory for the artifacts (bad ideas),
dereferences all links one by one after an
extract (quite complex) or simply dereferences
the links while doing the tarball (easiest way).

The latest solution has been used.