Arweave Configuration Redesign
Defines a new application called arweave_config designed
to deal with static and dynamic configuration. It should
also be compatible with legacy configuration and parameters.

The current way to pass the data from the entry-point
is a bit complex. Indeed, passing data to the BEAM can
only be done using rpc call or an evaluation. In both
case, the data should be sanitized and encoded first
to avoid single and double quotes issues. To do that,
the configuration should probably be encoded with
base64 (available on all system by default). If not,
an encoder will be required.

see: https://github.com/ArweaveTeam/arweave-dev/issues/940