Continue to work on configuration manage PoC
The current structure is a bit complex but should
do the job. The idea is to have two levels of
configuration.

The first level is about parameters specifications.
Those should be checked during compiled time. The
specifications are mainly used to deal with
arguments, environment variables and configuration
file parameters.

The second level is about arguments, environment
variables and configuration file. Those should be
checked during runtime (and startup). The function
to check them are from the specifications.