Added Environment PoC
While working on arguments, I was stuck because of the way we are
currently dealing with the entry point. The arguments are sent
to ar:module/1 and then parsed there. If arweave_config is a
dedicated application, a way to forward the arguments from
ar:module/1 will be required or the entry-point needs to be
moved to arweave_config.

Arweave_config_spec is becoming the main module/process to check
key/value and then forward them to arweave_config_store. The data
in arweave_config_store must be valid.

arweave_config_environment is storing a parsed snapshot of the
environment when the application is started. Those environment
can be loaded or displayed using arweave_config_environment:load/0
and arweave_config_environment:get/0.