Another strategy.
The amount of line to modify is insane and it can generate
lot of trouble. Instead of modifying everything and adding
new feature, I will use arweave_config:get/1 and
arweave_config:set/2 as interface to the old configuration.

Most of the old configuration still don't have any
namespaces or group attributed. The logic is quite different
and to avoid breaking the application, it should be
easier to do it step by step.