Rollback previous commit
The way the parser was wrong. While reading the existing invokation
example from Discord and via previous debugging session, we can
probably do something easier than what I was doing.

elements field can be removed. An argument can be a flag or an
argument, where an argument can ONLY have 1 value. It will make
things easier. If a value need to be splitted, the handler_set
or check function can do that.

type will define the kind of value passed to the argument, in our
case, boolean means a flag (it can also have an implicit value
when configured) or any kind of type defined in  the specificatoin.

The way data are stored can also be a bit weird. The namespace
should probably refer to an unique data store. For example,
if we want to modify peers configuration, we should deal with
the peer table, and not a shared table.