Revert to master
This commit does a full revert of all change in arweave application
regarding the new configuration method, except for ar module. While
doing any test wisth the test suite, it seems a part of the modification
is wrong but I don't know which one. Let assume for now only the
argument parsing needs to be supported.