fix: fix a race condition to can occur when syncing r29 data
The process for querying which r29 footprints are advertised by a peer can cause peer interval ranges to be enqueued which include chunks beyond the weave_size observed by this node (e.g. if the peer has applied more blocks than this node). This fix cuts those intervals so that they do not attempt to fetch chunks beyond the weave_size that this node has observed.