Integrate HTTP API Interface for Arweave Config
This commit integrate cowboy as part of arweave config to give access to
an HTTP API interface and dynamically configure parameters while arweave
is running.

The process in chage of this interface is not started automatically for
the moment, but can be started using two methods: (1) in "standalone"
mode using arweave_config_http_server:start_link() or (2) as child using
arweave_config_http_server:start_as_child(). The second method will start the
process under arweave_config_sup supervisor.

The API is using v0 prefix (this is not a stable interface for the
moment). Here the list of supported end-points:

  - /v0/config: export the whole configuration with parameter set

  - /v0/config/${parameter}: display a parameter value

  - /v0/environment: display the environment used by arweave

This implementation is using JSEND JSON format, instead of recreating
the wheel, JSEND format is easy and flexible enough for the private API
end-points.

Furthermore, to limit the attack surface and limit access to the
configuration HTTP API, this service can listen on an IP address OR an
Unix socket.

This commit also adds the support for multi-types. Indeed, a parameter
can have different type, for example a web server listen on an IP
address or an Unix socket. By default, the first matching type is used,
ignored the remaining ones.

Fixed a bug in arweave_config_spec:get/1.

Updated the documentation.

see: https://github.com/ArweaveTeam/arweave-dev/issues/978