Quick Fix: restart epmd in case of session leak
When using bin/start script, if an epmd session leaks
is found in the loop, the script should be able to restart
it. This procedure is only executed if ARWEAVE_EPMD_AUTO_RESTART
environment variable is set. Here other environment variable added:

 - ARWEAVE_RESTART_LIMIT
 - ARWEAVE_EPMD_AUTO_RESTART
 - ARWEAVE_EPMD_RESTART_METHOD

When arweave process is killed and ARWEAVE_EPMD_AUTO_RESTART environment
variable is set, the following message is returned:

    ./bin/start: line 212: 570756 Killed ${ARWEAVE} foreground ${*}
    Arweave Heartbeat: The Arweave server has been terminated by the system (SIGKILL).
    Arweave Heartbeat: Arweave will be restarted in 15 seconds.
    Arweave Heartbeat: If you would like to avoid this, press control+c to kill the server.
    Arweave Heartbeat: Start epmd restart procedure
    Arweave Heartbeat: epmd (554721) run as arweave:arweave with 0 sessions with ppid 1.

This feature is only useful when one server is dedicated to run one
arweave node. In case of many node running, the current procedure will
not work correctly. epmd is restart only if a session leak is detected.